Ship and ShipBase are separated
Ship is more of a wrapper for the base and the base is a couple of mostly empty objects.
With this logic, the base can be swapped out at any time and reducing the need for a lot of different prefabs that might need to be changed somewhere down the line.